requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-telegram-bot==21.6
SQLAlchemy==2.0.29
typing-inspection==0.4.2
typing_extensions==4.15.0
